pyproject.toml
[project]
name = "ai-quiz-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.19.0",
    "streamlit>=1.51.0",
    "google-generativeai",
    "matplotlib",
    "seaborn>=0.13.2",
    "pandas>=2.3.3",
    "numpy>=2.3.5",
    "google-genai>=1.52.0",
]
